requirements.txt
Django>=3.0.4,<3.1
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.6.0,<3.7.0
requests
